Refactor usage info parsing
